Brand the main FYSM page like the detail pages
Previously, the main FYSM page was branded like the "database" pages. This
is unfortunate, because we mostly don't want prefrosh to look at the rest
of the database, but do want them to look at the rest of the FYSM, so it
seems like it should match the FYSM instead of the rest of the "database".
Also, the navigation links to get to the rest of the "database" are now
gone.