Add infrastructure for per-group authz
This adds a PerGroupAuthz authentication backend (which is a no-op
for the authentication and only bothers with authorization).
OfficerRoles can specify users whose permissions they grant to the
holders of that role when working with that group. Those users can
then be granted the relevant permissions in the usual way.