Fix updater information
Systems updating entries should call set_updater with the updater (as a string
or request.user object) in order to ensure that the updater is correct.
Otherwise, it will be unset on instance.save(). Regardless, the update_time
will be updated.

This mechanism is independent of django-reversion's history functionality,
which also stores the updater.

Note, in particular, that the admin does not currently use this interface
to store the updater, but does store it in django-reversion's system (along
with the rest of the history). Porting the "view group" page over may be
desirable.