Allow non-Exec to actually edit groups
The StaticWidget would previously return the model as the value from a
foreign key field. This fails when it tries to get converted. Now, if we're
passed choices, we'll compare their labels against the unicode form of the
value, and see what matches.

Obviously, this means that your values need to unique. OTOH, (a) this is
probably a good idea, and (b) if you don't do this, your admin users, who
actually get to select a value, will have trouble selecting something
reasonable too. Consequently, I don't think I care that much.