Set cookie names
At the moment, we have several differet deployments of the ASA DB,
for various purposes. Two major ones are hosted on asa.mit.edu, at
/forms/ and /db2.0/. Unfortunately, use of the default cookie names
mean that you can't use the two sites simultaneously --- logging into
one will set the sessionid cookie to a site-specific value, which means
that the other site's value will no longer be there.

This allows setting a prefix for the cookie name. By setting this to
a different value for each deployment, we can avoid /forms/ logging out
/db2.0/, and vice-versa.