Register the models with reversion directly too
Instead of just using VersionAdmin to force models to be versioned,
register them directly in models.py. This has the advantage of meaning
that anywhere we use the models they'll be versioned.

IIRC, I'd tried this before and it didn't work, because I was registering
models more than once, but I can't remember the details of when it broke, and
I'm not seeing any breakage now. This may need to get reverted, though. (Or,
more likely, patched further to allow registering one model multiple times.)