Display full names of signatories
This displays the full name of signatories (not just the username) in the
group display page and the signatory change page (but not the big list
signatories page). It also requires a database query per signatory being
looked up. (This is one reason for not doing it on the big list signatories
page, the other being that full names take more space.)