Fix new groups in groups/diffs.py (Trac: #67)
This is three related fixes.

The Trac issue is that it turns out Django doesn't like negative queryset
indexes, so do a .reverse() and then a positive index...

The second issue is that, even if the negative indexed worked, we wanted to
be updating before, not after --- we want to set the original we're diffing
against, not the current version, despite the fact that we're grabbing the
original from after_versions instead of before_versions.

Finally, the third issue is that the comment about the no-change case is
misleading --- even when a group is just recognized, it still will have
multiple versions. In particular, the startup form submission + the approval.
This adds a clarifying comment.