Implement GroupFilter default values better
This solves a problem where if you subclass GroupFilter and add a required
field, the form will always be seen as bound, and consequently it will
give a validation error on initial load. We now use a "more right"
way to implement default values.

Also, as a convenience, if you pass in an empty data dictionary, we interpret
it as though it were none, so your form will not be bound if you use the
idiom of passing in request.GET regardless of whether your form has actually
been submitted.