Populate note author and group in "Add note" link
The group detail page has a link to add a new note about the group.
Historically, this just linked straight to the generic "Add group note"
admin page. That page takes GET parameters to prefill fields, so we now
submit the author and group information. It's certainly not a fix to
ASA Trac #77, but it does make it a slightly less desperate need.