Forbid commas in FYSM display names (ASA Trac #40)
This is perhaps overenthusiastic, especially since groups can't opt out. On
the other hand, we do actually have a handful of groups with commas that
aren't ", MIT" or ", The" that shouldn't have the comma in their display
name, and I sorta want to catch them too. Also, we have maybe one group
that should have a comma in their display name...