Use a new PAG for accessing constitutions
This fixes a potential privilege escalation issue from asa-internal (or anybody
else who can read constitutions) to asa-db-root. In particular, by putting a
path within the asa-db locker in for their constitution, they could convince
the DB to copy /mit/asa-db/.my.cnf or other sensitive files into the
constitutions directory, and then read it. By creating a new PAG, we drop the
daemon.scripts privileges and prevent the attack.

In a future change, we may wish to aklog with a new principal so as to be able
to read non-public constitutions. When we do so, we should be careful not to
use daemon.asa-db or any other principal with privileged read access to AFS.