Add newly created groups to asa-official (ASA-#111)
This fixes a bug added when groups/diffs.py was original created
(b6c7a4422c99475fee21445f136583e19751d838) by actually calling new_group when
appropriate. It uses the design from when asa-official updating was originally
added (4098256caab3ba58f03a31e194d705e766458570). It is reasonably likely that
we want to add groups to asa-official not when they get created but rather when
the become active, but that's a later change.