Unrequire description, website_url in admin
The only three fields that group members can edit (and thus that it makes
sense for Exec members to punt updating so that group members put correct
values in) are description, website_url, and officer_email. officer_email is
filled in for all but two (derecognized) groups, so it seems like it makes
sense to prevent Exec from accidentally unsetting it. description and
website_url are frequently not set, so unrequire them.