Serve membership/admin/issues.csv as text/csv
Firefox can be told to open a text/csv in the browser using the "Open in
browser" extension, so serving CSV files as text/plain no longer has
significant advantages (for me) over serving as text/csv, and has the
disadvantage of making it hard to open as a spreadsheet. Thus, I'm switching
back to text/csv.