Refer to "MIT ID number" in admin (ASA-#166)
Lemur points out that "card number" is a thing that means something different
from ID number, and that using it for "MIT ID number" in the
SpaceAccessListEntry admin is confusing for people who know that. So switch to
"MIT ID" (like we use elsewhere on the site), and add some help text for good
measure.