Diff asa-official with "correct" membership
This is a partial fix to ASA-#153. It reads the membership of asa-official and
compares it to the membership asa-official "should" have (in particular, the
listed officer_email for each group in the database with status active,
suspended, or nge). Deviations are reported, both in summary form and with the
individual email addresses.

Major remaining flaws:
* Support for the GSC and UA funding lists is half-written, and in particular
  requires that MoiraList's list_members implementation be finished
* The summary information should probably be printed at the top (for all
  lists), with the details below
* No output should be produced when there are no differences, so that it can
  be run from cron