Don't let people edit spaces the DB doesn't manage
Previously, this branch simply displayed a warning above Walker spaces, but
still allowed changing the access list. That seems a bit silly, so now we don't
display the access list update field at all.

The only group that had used the field for a Walker office is Camp Kesem
(https://asa.mit.edu/groups/1291/space/), so I also expired and notified the
people that they'd tried to grant access to.