Add help text about Athena lockername selection
As mentioned in ASA-#234, some people starting groups pick excessively long
locker names. This adds help text to try to alleviate that tendency (and also
to reduce the number of invalid characters people submit).

Unfortunately, enforcing this well seems difficult --- we want the SQL database
to support longer lockernames (some groups have org lockers instead of activity
lockers, and can therefore have longer names, for one), but have the frontend
(sometimes) reject it. It's not clear how to do this in Django.