Change "Update people" URL to say "people"
It had said "officers", which was inconsistent with the terminology used
elsewhere on that page.

This also adds a redirect from the old URL. The redirect view used is custom
and I *think* should work, but I'm not sure. Django doesn't appear to have
something built-in for this;
https://docs.djangoproject.com/en/dev/ref/class-based-views/base/#redirectview
doesn't seem to support view names.