Scripts auth: don't activate on 127.0.0.1 either
The scripts auth module has long delegated to the standard Django auth when the
hostname was localhost, in order to ignore local dev server instances. This
makes it also delegate to standard Django for 127.0.0.1 as well.  I'm not sure
why this hadn't come up before now... A quick look at the Django codebase
suggests this isn't a recent change.