People lookup: handle round numbers (ASA-#253)
It turns out that failure to split on commas isn't actually the cause of
ASA-#253 (though it is certainly a usability issue); instead, if we have 100*n
Athena usernames (most likely, zero), we end up with an empty "chunk" of
usernames to send to the LDAP server. That results in a filter like "(|)",
which errors. This changes the code to suppress empty chunks.